+++ noodles-gff/src/async.rs
@@ -1,9 +1,3 @@
 //! Async GFF.
 
 pub mod io;
-
-#[deprecated(
-    since = "0.33.0",
-    note = "Use `noodles_gff::r#async::io::Reader` instead."
-)]
-pub use self::io::Reader;